Imgui include oldui: Fix lightmaps in context menu (#4)
* dxvk_imgui.cpp - Lightmap/Hide Instance in UI

Added some hidden features to the UI, with a commented-out line for World Space UI Background.

* World Space Background, include Legacy UI in ImGui

Added the original deprecated UI as a toggle in ImGui and brought its features up to date with the modern UI.

* Old UI: Fix new filtered lists, hide leoleotte UI when active

When using the Legacy GUI, the new UI is now properly hidden. Lightmaps now instantly tag when clicked instead of showing a redundant context menu.